fix: thumb move flicker glitch seen on iOS
Ensure the event.target is self (range-slider) and not the thumb or any thumb children etc.

This avoids incorrect percent calculations when passing the event.offsetX/Y from a wrong target and not self. So far only seen on iOS devices on the initial touch move.

Might be a `pointer-events: none` issue on iOS.